+++ lib/src/modules/mixer.rs
@@ -1,51 +1,59 @@
-use alloc::vec::Vec;
 use alloc::vec;
+use alloc::vec::Vec;
 
 use graphity::signal::SignalGraph;
 use graphity::Node;
 
 use crate::core::signal::Signal;
 use crate::model::state::*;
-use crate::primitives;
 
-pub struct OscillatorNode {
-    oscillator: primitives::oscillator::Oscillator,
-    buffer: [f32; 32],
+pub struct MixerNode {
+    in1: [f32; 32],
+    in2: [f32; 32],
+    out: [f32; 32],
 }
 
 #[allow(clippy::new_without_default)]
-impl OscillatorNode {
+impl MixerNode {
     pub fn new() -> Self {
         Self {
-            oscillator: primitives::oscillator::Oscillator::new(48_000.0),
-            buffer: [0.0; 32],
+            in1: [0.0; 32],
+            in2: [0.0; 32],
+            out: [0.0; 32],
         }
     }
 }
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, Debug)]
-pub enum OscillatorConsumer {
-    Frequency,
+pub enum MixerConsumer {
+    In1,
+    In2,
 }
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, Debug)]
-pub struct OscillatorProducer;
+pub struct MixerProducer;
 
-impl Node<Signal> for OscillatorNode {
-    type Consumer = OscillatorConsumer;
-    type Producer = OscillatorProducer;
+impl Node<Signal> for MixerNode {
+    type Consumer = MixerConsumer;
+    type Producer = MixerProducer;
 
-    fn write(&mut self, _consumer: Self::Consumer, input: Signal) {
-        self.oscillator.frequency = input.as_control();
+    fn write(&mut self, consumer: Self::Consumer, input: Signal) {
+        match consumer {
+            MixerConsumer::In1 => self.in1 = input.as_audio(),
+            MixerConsumer::In2 => self.in2 = input.as_audio(),
+        }
     }
 
     fn read(&self, _producer: Self::Producer) -> Signal {
-        Signal::from_audio(self.buffer)
+        Signal::from_audio(self.out)
     }
 
     fn tick(&mut self) {
-        let buffer = &mut self.buffer;
-        self.oscillator.populate(buffer);
+        self.in1
+            .iter()
+            .zip(self.in2.iter())
+            .enumerate()
+            .for_each(|(i, (in1, in2))| self.out[i] = in1 + in2);
     }
 }
 
@@ -61,22 +69,27 @@ pub fn register<N, NI, CI, PI>(
     <NI as graphity::NodeIndex>::Consumer: From<crate::instrument::__Consumer>,
     <NI as graphity::NodeIndex>::Producer: From<crate::instrument::__Producer>,
 {
-    let oscillator = OscillatorNode::new();
+    let mixer = MixerNode::new();
 
     use crate::instrument::{__Consumer, __Node, __Producer};
-    let oscillator = graph.add_node::<__Node>(oscillator.into());
-    let consumer_frequency =
-        oscillator.consumer::<__Consumer>(OscillatorConsumer::Frequency.into());
-    let producer = oscillator.producer::<__Producer>(OscillatorProducer.into());
+    let mixer = graph.add_node::<__Node>(mixer.into());
+    let consumer_in1 = mixer.consumer::<__Consumer>(MixerConsumer::In1.into());
+    let consumer_in2 = mixer.consumer::<__Consumer>(MixerConsumer::In2.into());
+    let producer = mixer.producer::<__Producer>(MixerProducer.into());
 
     modules.push(Module {
-        handle: oscillator,
-        name: "OSC",
+        handle: mixer,
+        name: "MIX",
         index: 0,
         attributes: vec![
             Attribute {
-                socket: Socket::Consumer(consumer_frequency),
-                name: "FRQ",
+                socket: Socket::Consumer(consumer_in1),
+                name: "IN1",
+                connected: false,
+            },
+            Attribute {
+                socket: Socket::Consumer(consumer_in2),
+                name: "IN2",
                 connected: false,
             },
             Attribute {