+++ lib/src/modules/mod.rs
@@ -1,3 +1,4 @@
 pub mod audio_output;
 pub mod control_input;
+pub mod mixer;
 pub mod oscillator;